requirements.txt
python-dotenv
qdrant-client
langchain
langchain-community
langchain-google-genai
sentence-transformers
tiktoken
pypdf
huggingface-hub
langgraph
google-genai
fastapi
uvicorn


torch
torchvision
torchaudio
transformers